Changed camera to call intent instead of manual preview

+++ Shopport/app/src/main/java/nz/ac/uclive/shopport/database/WishListItem.kt
@@ -8,7 +8,7 @@ data class WishListItem (
     @PrimaryKey(autoGenerate = true) var id: Long = 0L,
     @ColumnInfo var title: String,
     @ColumnInfo var description: String,
-    @ColumnInfo var price: Double,
+    @ColumnInfo var price: Int,
     @ColumnInfo var location: String,
     @ColumnInfo var imageURI: String,
     @ColumnInfo var bought: Boolean = false)
